base de datos h2 configurada y andando
