--- src/main/java/com/musicapp/musicapp/dominio/Cancion.java
@@ -1,36 +1,0 @@
-package com.musicapp.musicapp.dominio;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.GenericGenerator;
-import org.hibernate.annotations.JdbcTypeCode;
-import org.hibernate.type.SqlTypes;
-
-import java.util.UUID;
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class Cancion {
-    @Id
-    @GeneratedValue(generator = "UUID")
-    @GenericGenerator(name="UUID")
-    @JdbcTypeCode(SqlTypes.CHAR)
-    @Column(length = 36, columnDefinition = "varchar(36)", updatable = false, nullable = false)
-    private UUID id;
-    @Column(nullable = false)
-    private String nombre;
-    private int ranking;
-    @Column(nullable = false)
-    private Double duracion;
-    private Genero genero;
-    @Column(nullable = false)
-    private Artista artista;
-    private String album;
-
-}
