--- src/main/java/com/musicapp/musicapp/dominio/Usuario.java
@@ -1,35 +1,0 @@
-package com.musicapp.musicapp.dominio;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.GenericGenerator;
-import org.hibernate.annotations.JdbcTypeCode;
-import org.hibernate.type.SqlTypes;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class Usuario extends Auditor {
-    @Id
-    @GeneratedValue(generator = "UUID")
-    @GenericGenerator(name="UUID")
-    @JdbcTypeCode(SqlTypes.CHAR)
-    @Column(length = 36, columnDefinition = "varchar(36)", updatable = false, nullable = false)
-    private UUID id;
-    @Column(nullable = false)
-    private String nombre;
-    @Column(nullable = false)
-    private String usuario;
-    private List<ListaDeReproduccion> listaDeReproduccion = new ArrayList<>();
-}
