+++ src/main/java/com/musicapp/musicapp/entity/Artista.java
@@ -1,0 +1,30 @@
+package com.musicapp.musicapp.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.UuidGenerator;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+@Entity
+@Table(name = "Artista")
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class Artista {
+    @Id
+//  https://www.baeldung.com/java-hibernate-uuid-primary-key
+//  @GeneratedValue(generator = "uuid-hibernate-generator")
+//  @GenericGenerator(name = "uuid-hibernate-generator", strategy = "org.hibernate.id.UUIDGenerator")
+    @UuidGenerator
+    private UUID id;
+    //@Column(nullable = false)
+    private String nombre;
+    @OneToMany(fetch = FetchType.EAGER, mappedBy = "artista", cascade = {CascadeType.PERSIST, CascadeType.REMOVE})
+    private List<Cancion> canciones = new ArrayList<>();
+}
