+++ src/main/java/com/musicapp/musicapp/entity/Cancion.java
@@ -1,0 +1,40 @@
+package com.musicapp.musicapp.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.UuidGenerator;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+@Entity
+@Table(name = "Cancion")
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class Cancion {
+    @Id
+    @UuidGenerator
+    private UUID id;
+    //@Column(nullable = false)
+    private String nombre;
+    private int ranking;
+    //@Column(nullable = false)
+    private Double duracion;
+    //@Column(nullable = false)
+    @ManyToOne
+    private Genero genero;
+    //@Column(nullable = false)
+    @ManyToOne
+    private Artista artista;
+    private String album;
+    @ManyToMany
+    @JoinTable(name = "listas_canciones", joinColumns = @JoinColumn(name = "cancion_nombre"),
+    inverseJoinColumns = @JoinColumn(name = "lista_id"))
+    private List<ListaDeReproduccion> listas = new ArrayList<>();
+
+}
