+++ src/main/java/com/musicapp/musicapp/entity/Usuario.java
@@ -1,0 +1,29 @@
+package com.musicapp.musicapp.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.UuidGenerator;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+@Entity
+@Table(name = "Usuario")
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class Usuario extends Auditor {
+    @Id
+    @UuidGenerator
+    private UUID id;
+    //@Column(nullable = false)
+    private String nombre;
+    //@Column(nullable = false)
+    private String usuario;
+    @OneToMany(fetch = FetchType.EAGER, mappedBy = "usuario", cascade = {CascadeType.PERSIST, CascadeType.REMOVE})
+    private List<ListaDeReproduccion> listaDeReproduccion = new ArrayList<>();
+}
