+++ src/main/resources/application.properties
@@ -1,1 +1,9 @@
-
+server.port = 8080
+spring.datasource.url = jdbc:h2:mem:testdb
+spring.datasource.driver-class-name = org.h2.Driver
+spring.datasource.username = sa
+spring.datasource.password = sa
+spring.h2.console.enabled = true
+spring.jpa.database-platform = org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto = create
+spring.jpa.show-sql = true
